backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.0
openpyxl==3.1.2
numpy==1.26.4
opencv-python==4.6.0.66
python-dotenv==1.0.0
setuptools==70.0.0
paddlepaddle==3.2.2
paddleocr==3.3.2